// src/enums/profession.enum.ts
export enum ProfessionEnum {
  AGRICULTURE_ENVIRONMENT = "AGRICULTURE_ENVIRONMENT",
  REAL_ESTATE_CONSTRUCTION = "REAL_ESTATE_CONSTRUCTION",
  INDUSTRY_ENERGY = "INDUSTRY_ENERGY",
  DIGITAL_TELECOM = "DIGITAL_TELECOM",
  RETAIL_DISTRIBUTION = "RETAIL_DISTRIBUTION",
  TRANSPORT_LOGISTICS = "TRANSPORT_LOGISTICS",
  FINANCE_INSURANCE = "FINANCE_INSURANCE",
  HEALTH_SOCIAL = "HEALTH_SOCIAL",
  EDUCATION_RESEARCH = "EDUCATION_RESEARCH",
  PUBLIC_ADMINISTRATION = "PUBLIC_ADMINISTRATION",
  MILITARY_SECURITY = "MILITARY_SECURITY",
  CATERING_TOURISM = "CATERING_TOURISM",
  CULTURE_LEISURE = "CULTURE_LEISURE",
  ARTS_MEDIA = "ARTS_MEDIA",
  OTHER_SECTOR = "OTHER_SECTOR",
}